+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.java
@@ -32,10 +32,7 @@ import at.hannibal2.skyhanni.features.garden.composter.ComposterDisplay;
 import at.hannibal2.skyhanni.features.garden.composter.ComposterInventoryNumbers;
 import at.hannibal2.skyhanni.features.garden.composter.GardenComposterInventoryFeatures;
 import at.hannibal2.skyhanni.features.garden.farming.*;
-import at.hannibal2.skyhanni.features.garden.inventory.GardenDeskInSBMenu;
-import at.hannibal2.skyhanni.features.garden.inventory.GardenInventoryNumbers;
-import at.hannibal2.skyhanni.features.garden.inventory.GardenNextPlotPrice;
-import at.hannibal2.skyhanni.features.garden.inventory.SkyMartCopperPrice;
+import at.hannibal2.skyhanni.features.garden.inventory.*;
 import at.hannibal2.skyhanni.features.garden.visitor.GardenVisitorColorNames;
 import at.hannibal2.skyhanni.features.garden.visitor.GardenVisitorFeatures;
 import at.hannibal2.skyhanni.features.garden.visitor.GardenVisitorTimer;
@@ -271,6 +268,7 @@ public class SkyHanniMod {
         loadModule(new ToolTooltipTweaks());
         loadModule(new CropSpeedMeter());
         loadModule(new AshfangMinisNametagHider());
+        loadModule(new GardenTeleportPadInventoryNumber());
 
         Commands.INSTANCE.init();